+++ app/Http/Controllers/Jihans/GudangReturnController.php
@@ -46,7 +46,7 @@ class GudangReturnController extends Controller
                 ->where('jihans_retail_stock.quantity', '>', 0)
                 ->select('master_products.*', 'jihans_retail_stock.quantity as current_stock')
                 ->with('unit')->orderBy('master_products.name')->get()
-                ->map(fn ($p) => ['id' => $p->id, 'name' => $p->name, 'code' => $p->code, 'unit_id' => $p->unit_id, 'unit_name' => $p->unit?->abbreviation ?? 'PCS', 'stock' => (float) $p->current_stock]),
+                ->map(fn ($p) => ['id' => $p->id, 'name' => $p->name, 'code' => $p->code, 'unit_id' => $p->unit_id, 'unit_name' => $p->unit?->abbreviation ?? 'PCS', 'stock' => (int) $p->current_stock]),
             'units'    => Unit::orderBy('name')->get()->map(fn ($u) => ['id' => $u->id, 'abbreviation' => $u->abbreviation]),
         ]);
     }
