+++ app/Http/Resources/Gudang/TransferRequestResource.php
@@ -35,10 +35,10 @@ class TransferRequestResource extends JsonResource
                 'product'           => $d->product?->name ?? '-',
                 'product_code'      => $d->product?->code ?? '-',
                 'unit'              => $d->unit?->abbreviation ?? 'PCS',
-                'quantity_requested' => (float) $d->quantity_requested,
-                'quantity_approved' => $d->quantity_approved !== null ? (float) $d->quantity_approved : null,
+                'quantity_requested' => (int) $d->quantity_requested,
+                'quantity_approved' => $d->quantity_approved !== null ? (int) $d->quantity_approved : null,
                 // Attached in the controller to avoid an N+1 stock lookup per row.
-                'warehouse_stock'   => (float) ($d->warehouse_stock ?? 0),
+                'warehouse_stock'   => (int) ($d->warehouse_stock ?? 0),
             ])),
         ];
     }
